requirements.txt
aniso8601==8.0.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-RESTful==0.3.8
frozendict==1.2
future==0.18.2
geocoder==1.38.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
probableparsing==0.0.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-crfsuite==0.9.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
ratelim==0.1.6
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
typing==3.7.4.1
urllib3==1.25.8
usaddress==0.5.10
usaddress-scourgify==0.1.11
waitress==1.4.3
Werkzeug==1.0.0
yaml-config==0.1.4
